code quality improvements
