+++ apps/web/modules/survey/multi-language-surveys/lib/types.ts
@@ -1,0 +1,24 @@
+import { TI18nString } from "@formbricks/types/i18n";
+
+export interface TranslatableString {
+  path: string;
+  displayId: string;
+  fieldLabel: string;
+  value: TI18nString;
+  isRichText: boolean;
+  elementId: string;
+}
+
+export interface TranslationProgress {
+  translated: number;
+  total: number;
+  percentage: number;
+}
+
+export interface DeduplicatedString {
+  key: string;
+  displayId: string;
+  fieldLabel: string;
+  value: TI18nString;
+  duplicatePaths: string[];
+}
